+++ src/elements/alert-dialog/alert-dialog.tsx
@@ -1,32 +1,156 @@
 'use client'
 
-import type { FC, ReactNode } from 'react'
+import {
+  Action,
+  Cancel,
+  Content,
+  Description,
+  Overlay,
+  Portal,
+  Root,
+  Title,
+  Trigger
+} from '@radix-ui/react-alert-dialog'
+import clsx from 'clsx'
+import {
+  type ComponentPropsWithoutRef,
+  type ElementRef,
+  type FC,
+  type HTMLAttributes,
+  type ReactNode,
+  forwardRef
+} from 'react'
 
 import { useDictionary } from '@/contexts/root/dictionary-provider'
 
-import {
-  _AlertDialogAction,
-  _AlertDialogCancel,
-  _AlertDialogContent,
-  _AlertDialogDescription,
-  _AlertDialogFooter,
-  _AlertDialogHeader,
-  _AlertDialogRoot,
-  _AlertDialogTitle,
-  _AlertDialogTrigger
-} from './_alert-dialog'
+const AlertDialogRoot = Root
 
-interface IAlertDialogBodyProps {
+const AlertDialogTrigger = Trigger
+
+const AlertDialogPortal = Portal
+
+const AlertDialogOverlay = forwardRef<
+  ElementRef<typeof Overlay>,
+  ComponentPropsWithoutRef<typeof Overlay>
+>(({ className, ...props }, ref) => (
+  <Overlay
+    className={clsx(
+      'fixed inset-0 z-50 bg-black/80 data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0',
+      className
+    )}
+    {...props}
+    ref={ref}
+  />
+))
+AlertDialogOverlay.displayName = Overlay.displayName
+
+const AlertDialogContent = forwardRef<
+  ElementRef<typeof Content>,
+  ComponentPropsWithoutRef<typeof Content>
+>(({ className, ...props }, ref) => (
+  <AlertDialogPortal>
+    <AlertDialogOverlay />
+    <Content
+      ref={ref}
+      className={clsx(
+        'fixed left-[50%] top-[50%] z-50 grid w-full max-w-lg translate-x-[-50%] translate-y-[-50%] gap-4 border bg-background p-6 shadow-lg duration-200 data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[state=closed]:slide-out-to-left-1/2 data-[state=closed]:slide-out-to-top-[48%] data-[state=open]:slide-in-from-left-1/2 data-[state=open]:slide-in-from-top-[48%] sm:rounded-lg',
+        className
+      )}
+      {...props}
+    />
+  </AlertDialogPortal>
+))
+AlertDialogContent.displayName = Content.displayName
+
+const AlertDialogHeader = ({
+  className,
+  ...props
+}: HTMLAttributes<HTMLDivElement>) => (
+  <div
+    className={clsx(
+      'flex flex-col space-y-2 text-center sm:text-left',
+      className
+    )}
+    {...props}
+  />
+)
+AlertDialogHeader.displayName = 'AlertDialogHeader'
+
+const AlertDialogFooter = ({
+  className,
+  ...props
+}: HTMLAttributes<HTMLDivElement>) => (
+  <div
+    className={clsx(
+      'flex flex-col-reverse sm:flex-row sm:justify-end sm:space-x-2',
+      className
+    )}
+    {...props}
+  />
+)
+AlertDialogFooter.displayName = 'AlertDialogFooter'
+
+const AlertDialogTitle = forwardRef<
+  ElementRef<typeof Title>,
+  ComponentPropsWithoutRef<typeof Title>
+>(({ className, ...props }, ref) => (
+  <Title
+    ref={ref}
+    className={clsx('text-lg font-semibold', className)}
+    {...props}
+  />
+))
+AlertDialogTitle.displayName = Title.displayName
+
+const AlertDialogDescription = forwardRef<
+  ElementRef<typeof Description>,
+  ComponentPropsWithoutRef<typeof Description>
+>(({ className, ...props }, ref) => (
+  <Description
+    ref={ref}
+    className={clsx('text-sm text-muted-foreground', className)}
+    {...props}
+  />
+))
+AlertDialogDescription.displayName = Description.displayName
+
+const AlertDialogAction = forwardRef<
+  ElementRef<typeof Action>,
+  ComponentPropsWithoutRef<typeof Action>
+>(({ className, ...props }, ref) => (
+  <Action
+    ref={ref}
+    className={clsx('button variant--primary size--md', className)}
+    {...props}
+  />
+))
+AlertDialogAction.displayName = Action.displayName
+
+const AlertDialogCancel = forwardRef<
+  ElementRef<typeof Cancel>,
+  ComponentPropsWithoutRef<typeof Cancel>
+>(({ className, ...props }, ref) => (
+  <Cancel
+    ref={ref}
+    className={clsx('button variant--outline size--md mt-2 sm:mt-0', className)}
+    {...props}
+  />
+))
+AlertDialogCancel.displayName = Cancel.displayName
+
+interface IAlertDialogProps {
   title: string
   description: string
+  trigger: ReactNode
   cancelLabel?: string
   actionLabel?: string
   children?: ReactNode
 }
 
-const AlertDialogBody: FC<IAlertDialogBodyProps> = ({
+export const AlertDialog: FC<IAlertDialogProps> = ({
   title,
   description,
+  trigger,
   cancelLabel,
   actionLabel,
   children
@@ -34,27 +158,25 @@ const AlertDialogBody: FC<IAlertDialogBodyProps> = ({
   const { dictionary } = useDictionary()
 
   return (
-    <_AlertDialogContent>
-      <_AlertDialogHeader>
-        <_AlertDialogTitle>{title}</_AlertDialogTitle>
-        <_AlertDialogDescription>{description}</_AlertDialogDescription>
-        {children}
-      </_AlertDialogHeader>
-      <_AlertDialogFooter>
-        <_AlertDialogCancel>
-          {cancelLabel ?? dictionary.DIALOG_BUTTON.CANCEL}
-        </_AlertDialogCancel>
-        <_AlertDialogAction>
-          {actionLabel ?? dictionary.DIALOG_BUTTON.CONTINUE}
-        </_AlertDialogAction>
-      </_AlertDialogFooter>
-    </_AlertDialogContent>
+    <AlertDialogRoot>
+      <AlertDialogTrigger asChild>{trigger}</AlertDialogTrigger>
+
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>{title}</AlertDialogTitle>
+          <AlertDialogDescription>{description}</AlertDialogDescription>
+          {children}
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>
+            {cancelLabel ?? dictionary.DIALOG_BUTTON.CANCEL}
+          </AlertDialogCancel>
+          <AlertDialogAction>
+            {actionLabel ?? dictionary.DIALOG_BUTTON.CONTINUE}
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialogRoot>
   )
 }
-AlertDialogBody.displayName = 'AlertDialogBody'
-
-export const AlertDialog = {
-  Root: _AlertDialogRoot,
-  Trigger: _AlertDialogTrigger,
-  Body: AlertDialogBody
-}
+AlertDialog.displayName = 'AlertDialog'
