+++ src/elements/collapsible/collapsible.tsx
@@ -5,6 +5,7 @@ import {
   CollapsibleTrigger as RadixCollapsibleTrigger,
   Root
 } from '@radix-ui/react-collapsible'
+import type { FC, ReactNode } from 'react'
 
 const CollapsibleRoot = Root
 
@@ -12,8 +13,17 @@ const CollapsibleTrigger = RadixCollapsibleTrigger
 
 const CollapsibleContent = RadixCollapsibleContent
 
-export const Collapsible = {
-  Root: CollapsibleRoot,
-  Trigger: CollapsibleTrigger,
-  Content: CollapsibleContent
+interface ICollapsibleProps {
+  trigger: ReactNode
+  content: ReactNode
 }
+
+export const Collapsible: FC<ICollapsibleProps> = ({ trigger, content }) => {
+  return (
+    <CollapsibleRoot>
+      <CollapsibleTrigger asChild>{trigger}</CollapsibleTrigger>
+      <CollapsibleContent>{content}</CollapsibleContent>
+    </CollapsibleRoot>
+  )
+}
+Collapsible.displayName = 'Collapsible'
